move health server into separated package
